bibliothecaire_service/requirements.txt
Flask
python-dotenv
PyJWT
cryptography
requests
flask-pymongo
flask-restx
bcrypt